# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API BASE",
    "version": "1.0.0",
    "description": "API de base pour la gestion de l'authentification et des logs.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Mr NJIFANDA",
      "url": "https://njifanda.com",
      "email": "contact@njifanda.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Base - API (mode: Development)"
    }
  ],
  "paths": {
...
    "/app/upload/upload": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload single file",
        "description": "Upload a single file to the server",
        "operationId": "uploadFile",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional metadata for the file",
                    "properties": {
                      "description": {
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "File uploaded successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "filename": {
                          "type": "string",
                          "example": "file-1234567890.jpg"
                        },
                        "originalname": {
                          "type": "string",
                          "example": "photo.jpg"
                        },
                        "mimetype": {
                          "type": "string",
                          "example": "image/jpeg"
                        },
                        "size": {
                          "type": "integer",
                          "example": 102400
                        },
                        "path": {
                          "type": "string",
                          "example": "/uploads/file-1234567890.jpg"
                        },
                        "uploaded_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/app/upload/upload-multi": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload multiple files",
        "description": "Upload multiple files at once",
        "operationId": "uploadMultipleFiles",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of files to upload"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional metadata for all files",
                    "properties": {
                      "description": {
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Files uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Files uploaded successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "originalname": {
                            "type": "string"
                          },
                          "mimetype": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "path": {
                            "type": "string"
                          },
                          "uploaded_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/app/upload/stats": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "Get upload statistics",
        "description": "Retrieve statistics about uploaded files",
        "operationId": "getUploadStats",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date for statistics (ISO 8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date for statistics (ISO 8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by specific user",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_files": {
                          "type": "integer",
                          "example": 150
                        },
                        "total_size": {
                          "type": "integer",
                          "example": 52428800
                        },
                        "files_by_type": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "integer"
                          },
                          "example": {
                            "image/jpeg": 50,
                            "application/pdf": 30
                          }
                        },
                        "uploads_today": {
                          "type": "integer",
                          "example": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/app/upload/delete": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Delete uploaded file",
        "description": "Delete a previously uploaded file",
        "operationId": "deleteUploadedFile",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "file_id"
                ],
                "properties": {
                  "file_id": {
                    "type": "string",
                    "description": "File ID to delete",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/app/upload/info": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "Get file information",
        "description": "Retrieve detailed information about an uploaded file",
        "operationId": "getFileInfo",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "query",
            "required": true,
            "description": "File ID",
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "filename": {
                          "type": "string"
                        },
                        "originalname": {
                          "type": "string"
                        },
                        "mimetype": {
                          "type": "string"
                        },
                        "size": {
                          "type": "integer"
                        },
                        "path": {
                          "type": "string"
                        },
                        "uploader": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        },
                        "metadata": {
                          "type": "object"
                        },
                        "uploaded_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key required for all endpoints"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated endpoints"
      }
    },
    "responses": {
      "SuccessResponse": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Operation successful"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "BadRequestError": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Validation failed"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
}